requirements.txt
tensorflow
matplotlib
scikit-learn
mlflow
dagshub
python-dotenv
pytest
deepchecks[vision]
scikit-image
torch
torchvision
fastapi
uvicorn